YUK-14563 Data Export Null Pointer Exception
Adding intValue for getDaysPrevious as per Stacey's suggestion.

+++ yukon-client/common/src/com/cannontech/amr/archivedValueExporter/dao/impl/ArchiveValuesExportFormatDaoImpl.java
@@ -216,10 +216,10 @@ public class ArchiveValuesExportFormatDaoImpl implements ArchiveValuesExportForm
         
         for (ExportField field:format.getFields()) {
             
-            if (field.getField().getType().equals(FieldType.ATTRIBUTE) 
-                    && field.getField().getAttribute().getAttribute() == attribute.getAttribute()
-                    && field.getField().getAttribute().getDataSelection() == attribute.getDataSelection()
-                    && field.getField().getAttribute().getDaysPrevious() == attribute.getDaysPrevious()) {
+            if (field.getField().getType().equals(FieldType.ATTRIBUTE)
+                && field.getField().getAttribute().getAttribute() == attribute.getAttribute()
+                && field.getField().getAttribute().getDataSelection() == attribute.getDataSelection()
+                && field.getField().getAttribute().getDaysPrevious().intValue() == attribute.getDaysPrevious().intValue()) {
                 exportFields.add(field);
             }
         }
